test: Get rid of 4.9 pipeline
In v1.13 we stopped supporting the 4.9 kernel [1]. Therefore, we can
remove the 4.9 CI pipeline.

Previously, the 4.9 CI used to run with kube-proxy and KPR=disabled,
while the 4.19 CI with kube-proxy and KPR=strict (aka mixed mode). This
commit sets KPR=disabled in the 4.19 CI. The mixed mode is tested in the
new ci-datapath GHA (minus the IPv6 until [2] has been resolved; this is
fine).

[1]: https://github.com/cilium/cilium/issues/22116
[2]: https://github.com/cilium/cilium-cli/issues/1139

Signed-off-by: Martynas Pumputis <m@lambda.lt>