slb: cec nodeselector for t2 nodes
Currently, it's expected that Envoy is only installed on the
SLB T2 nodes by configuring the Cilium Helm Chart accordingly
(Helm value `envoy.nodeSelector`).

In addition to this, this commit also configures the nodeselector
on the CiliumEnvoyConfig correctly so that the config only gets applied
on T2 nodes, even if Envoy would be configured on T1 nodes.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>