slb: replace inline routeconfig with RDS
Currently, the Envoy RouteConfiguration is inlined in the definition
of every Envoy Listener. In terms of scalability (e.g. only sending
changed Envoy resources over the wire via xDS) it's better to use
RDS (RouteDiscoveryService) and define the RouteConfiguration separately
from the Listener.

Therefore, this commit changes the logic that the RouteConfiguration of
every IsovalentLB gets distributed via RDS instead of having it inline.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>