slb: set frontend port on endpoint port
Currently the T1 endpoint port is hardcoded to port `80`.

Even though everything is working as Cilium doesn't use this port,
this commit changes the port to be the same as the frontend port that
is also used as service port.

Otherwise this might be confusing to users who are looking at the Endpoints.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>