slb: don't merge annotations and labels when updating owned resources
Currently, when updating resources owned by the Standalone LB controlplane
(`Service`, `Endpoints` & `CiliumEnvoyConfig`), the desired labels and annotations
are merged with the existing ones. Initial thought was to allow manual
modification of annotations and labels on these resources.

But this comes with the drawback that reverting changes (e.g. removing
LB IPAM annotations) wouldn't be possible without additional logic.

As we don't make use of this merge functionality as of now (and probably
want to have control over these annotations and labels anyway), this comit
removes this functionality.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>