slb: initial support for assigning dynamic VIP from LB IPAM pool
Currently, we only have support for static frontend VIP set on `IsovalentLB.spec.vip`.

We still rely on the presence of an existing `CiliumLoadBalancerIPPool` but we
always set the statically defined VIP via the LB IPAM annotation
`lbipam.cilium.io/ips` on the T1 Service. In addition the static VIP also gets
configured on the T2 Envoy listener via `CiliumEnvoyConfig`.

This commit adds support for dynamic LB IPAM IP allocation. If no VIP
is defined on the `IsovalentLB.spec.vip` (empty string), the LB IPAM
annotation isn't added to the T1 Service. In addition to this, the reconciliation
logic has been refactored to await the creation of the T1 Service and the assignment
of an loadbalancer IP via LB IPAM before creating the T2 CiliumEnvoyConfig. This
is necessary because the Envoy Listener gets configured with the same VIP and port as
the T1 Service.

The k8s watchers on owned Services ensure that the reconciliation logic is triggered
after LB IPAM updated the status `Service.status.loadBalancer.ingress.ip`.

Note: the following aspects aren't part of this commit.

- CRD status handling (report assigned IP)
- CRD design (e.g. reference VIP pool, having VIP as optional value)
- LB IPAM IP sharing is only enabled in case of a static VIP. Sharing dynamic VIPs needs
  to become a concept of the LB CRDs if we want to support that.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>