slb: introduce initial unit-tests for CRD translation
This commit introduces a unit-test to test the translation between
the `IsovalentLB` CRD and the `Service`, `Endpoints` & `CiliumEnvoyConfig`.

The idea is to define the input and expected output (CRD definitions) as golden files
in the directory `testdata`. One directory with the YAMLs represents one testcase.

The testlogic itself loads all testcases, executes the translation logic and
compares the desired state with the exepcted one.

Upcoming features (and their translations) only have to be added as additional testcases.
Without the need to write Go code and nested Go/Proto structures.

Note: These tests doesn't remove the need for an actual integration test of the
reconciliation (where interaction with the k8s client and setting the ownerreference
is tested.)

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>