slb: add tls configuration to LBFrontend
This commit introduces the ability to configure TLS certificates
on the Standalone LB CRD `LBFrontend`.

```
apiVersion: isovalent.com/v1alpha1
kind: LBFrontend
metadata:
  name: lb-1
spec:
  port: 80
  tls:
    domainnames:
      - "*.bar.com"
      - "foo.acme.io"
    certificates:
      - secretName: "test"
      - secretName: "test2"
  ...
  routes:
  ...
```

This provides the possibility to configure TLS certificates as a list of references to K8s Secrets
of type TLS within the same namespace as the `LBFrontend` resource itself.

In addition, the domainnames covered by all certificates needs to be provided as a list of domainnames.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>