slb: support HTTPS route
This commit adds support for HTTPS on the LBFrontend. If TLS is configured on
the LBFrontend, every HTTP route is handled as HTTPS route too.

The implementation addsd a second Filter to the Envoy Listener FilterChain.

- FilterChain match "raw_buffer" -> HTTP (with HealthCheckFilter)
- FilterChain match "tls" on the configured domain names -> HTTPs

Once we're able to configure hostnames on the HTTP route, we're able to fine-tune
the configuration (without exposing everything on HTTP too).

Note: The HealthCheckFilter (for T1->T2 health checks) is only applied on the HTTP
filter, as T1 doesn't support HTTPS.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>