slb: use prefix 'lbfe-' as name of owned resources
Currently the name of created and owned resources (`Service`,
`Endpoints` & `CiliumEnvoyConfig`) are the same as the name of the
owning `LBFrontend`.

To prevent clashes with other functionality (other Cilium modules
or if we would have the need for separate `CiliumEnvoyConfig` for other
CRDs (e.g. Envoy Clusters in `CiliumEnvoyConfig` for `LBBackend`s) it's
better to prefix all names of created resources with `lbfe-`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>