slb: configure Envoy BPFMetadata listener filter with BPF map lookups disabled
Currently, the Cilium BPFmetadata listener is automatically added to every Envoy
listener that is configured via a `CiliumEnvoyConfig` - if the listener filter is
not already present in the listeners' config.

This results in unnecessary BPF map lookups and warnings (policy not found), as
the information that are looked up from the BPF maps aren't used by any Cilium
policy filters (because they aren't configured).

Nevertheless it's still necessary to add the listener filter because it's also
responsible to set some socket options that require elevated privileges.

Therefore, this commit adds an explicit BPF Metadata listener filter that disables
the bpf map lookups by setting `bpf_root` to an empty string.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>