slb: restrict T2 health check filter to T1 requests
Currently, the health check filter intercepts and responds to HTTP requests
on the actual VIP&Port if the method is `GET` and path is `/health`.

This might lead to problems if a user wants to loadbalance the same HTTP requests.

Therefore, this commit adds an additional match based on the user-agent that
needs to start with `cilium-probe/` (set by T1 to `cilium-probe/<version>`).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>