slb: remove need for cleanup scripts in examples
Currently there's a dedicated example cleanup script that removes the
docker containers that are attached to the same docker network as the
kind cluster (frr client & test backends).

This commit removes this need by using the improved version of kind-down
that supports deleting this containers in the same deletion process.

This functionality is in the process of getting upstreamed
(https://github.com/cilium/cilium/pull/33591), but we add it here so
we can already get rid of the scripts.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>