lb: removed all occurances of "standalone" in config flags
This commit removes all occurances of "standalone" in the config flags.

In most of the cases `standalone-lb-*` gets replaced with
`loadbalancer-cp-` (or LoadBalancer control plane) to avoid confusion
with some `loadbalancer-*` config flags that historically don't have
enough context. In addition i tried to be as expressive with the
descriptions of the flags as possible.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>