lb: introduce httpsroute on lbfrontend
This commit introduces the httpsroute on the LBFrontend CRD and moves
the configuration of domainNames from the root of LBFrontend to the routes
(`hostNames`).

This way, different domains can be treated by either HTTP and/or HTTPS on
the same LB frontend.

A LBFrontend with only a https route still configures a http filterchain entry
to serve the healthchecks (without any actual routes attached).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>