lb: replace hardcoded T1->T2 HC interval
This commit replaces the hardcoded interval (5s) that is used for health
checking between T1 and T2 with an interval that is half of the shortest
interval of all backends that belong to the frontend. This way we shouldn't
have a delay when propagating the HC information from T2 to T1.

In addition, the quarantine timeout (default 30s) has been disabled.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>