slb: Introduce LBVIP reconciler
Introduce a controller for LBVIP resource. Users can either request the
specific IPv4 VIP by setting LBVIP.spec.ipv4Request explicitly or let
Cilium Operator to pick up one from the pool. This controller supports
both operations.

The underlying implementation leverages an existing LBIPAM. Therefore
claims VIPs through Service. This Service is a kind of placeholder which
doesn't have any endpoints and only exists for claiming the VIP and
sharing the VIP with the T1 Services with sharing-key annotation.

One issue for this implementation is we must reserve at least one port
for this placeholder Service and T1 Services cannot use it. We chose
port number 1 as a fixed value. This is a reserved port for TCPMUX which
is a historic protocol and is not widely used anymore. Hopefully no one
uses it. We can get rid of this problem by removing LBIPAM dependency in
the future.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>