lb: use cluster discovery type EDS
Currently, we configure the Envoy clusters to use discovery type
`STRICT` to resolve the endpoints of the cluster.

Envoy recommends to use discovery type `EDS` in addition to `CDS` for
the following reason:

```
Although it is possible to use CDS without EDS by specifying fully static clusters, we recommend still using the EDS API for clusters specified via CDS. Internally, when a cluster definition is updated, the operation is graceful. However, all existing connection pools will be drained and reconnected. EDS does not suffer from this limitation. When hosts are added and removed via EDS, the existing hosts in the cluster are unaffected.
```

https://www.envoyproxy.io/docs/envoy/latest/intro/arch_overview/operations/dynamic_configuration#cds

Therefore, this commit introduces endpoint discovery via `EDS`.

We still have to revisit this once we enable DNS based endpoint resolution.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>