lb: normalize http path according to RFC 3986
This commit introduces HTTP path normalization according to RFC 3986.

```
Should paths be normalized according to RFC 3986 before any processing of requests by HTTP filters or routing? This affects the upstream :path header as well. For paths that fail this check, Envoy will respond with 400 to paths that are malformed. This defaults to false currently but will default true in the future. When not specified, this value may be overridden by the runtime variable http_connection_manager.normalize_path. See Normalization and Comparison for details of normalization. Note that Envoy does not perform case normalization
```

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>