lb: explicitly configure T2 HC no_traffic_interval
Currently, we don't configure `no_traffic_interval` on T2 Envoy
health check configuration explicitly.

Even though we configure the rest of the intervals according to the
values from the `LBBackend`, the health check interval defaults to `60s`
if the cluster doesn't receive any traffic.

To avoid confusion, this commit configures the `no_traffic_interval` with
the same value as all other intervals.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>