lb: don't execute loadbalancer e2e tests with go test ./...
Currently when executing `go test ./...` locally, the Loadbalancer E2E tests
are executed and fail due to the fact that the LB isn't set up.

In CI (`/ci-integration`) the problem doesn't appear as we already skip the
tests if `INTEGRATION_TESTS=true`.

Therefore, this commit changes the skip test to check for the  environment variable
`LOADBALANCER_TESTS` to be set to `true` - otherwise it gets skipped.

Note: Go build tags would have been an alternative approach. But this results in
issues with the IDE if the tags aren't set.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>