lb: Enable use_remote_address for HTTP proxies
Enable use_remote_address option of HTTP Connection Manager. It should
be enabled when the downstream is "untrusted" which means when we're not
sure the x-forwarded-for from the downstream is accurate or not. In our
case, we cannot make any assumption about who the downstream is, so we
should enable it.

This makes Envoy to append the downstream address to x-forwarded-for
header by default. If we want to disable it we still can toggle it using
skip_xff_append option. We can provide an interface to opt-out the XFF
later if needed.

Also, we remove x-envoy-internal and x-envoy-external-address headers
which are related to use_remote_address. While they are useful, we'll
remove it because we don't want users to rely on the envoy-specific
behavior.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>