lb: configurable tcp listener accesslog
While LB controlplane already supports configuring the Envoy access log
for the HTTPConnectionManager (HTTP(S)Proxy) and TCPProxy (TLSPassthrough),
this commit adds support for configuring the access log (format) for the TCP
listener itself.

This is especially useful to detect issues during the TLS handshake that are
reported with the access log operator `%DOWNSTREAM_TRANSPORT_FAILURE_REASON%`
(that is also added to the default format).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>