lb: extra config flag to enable tcp listener access log
With the enabled access log on the TCP listener, every successful
request results in two access log entries (one from the TCP listener and
one from the HCM/TCPProxy). In addition,  the TCP listener also outputs access
log entries for all T1->T2 health check requests - because it's not possible
to filter these requests based on HTTP headers the same way we do this on the HCM
access log.

To reduce noice, but still support the accesslog access log with a reasonable default
format for TLS debugging purposes, this commit introduces a new config flag
`loadbalancer-cp-accesslog-enable-tcp` of type `bool` to enable the TCP listener
access log explicitly on demand. It's disabled by default.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>