lb: fix lb ipam ip sharing
Currently, when changing the VIP of a `LBFrontend` (e.g. by changing
LBVIP reference), LB IPAM doesn't change the IP of the underlying k8s
Service correctly (to actually be the same of the k8s Service of the LBVIP).

The reason is that the sharing key annotation alone isn't enough if both
Services already have an IP assigned by LB IPAM. In this case LB IPAM
actually doesn't know what IP to share.

Therefore, with this commit, the k8s Service of an `LBFrontend` always
explicitly requests the same IP that the LBVIP got assigned by LB IPAM, by
always setting the assigned IP in the annotation `lbipam.cilium.io/ips`.

This way, whenever the VIP changes, the underlying services are updated correctly.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>