lb: Reject requests with underscore in the header field
This is required for protecting the CGI applications that maps hyphens
to underscores on request and underscores to hyphens on response. Such
applications treats hyphens and underscores interchangeable and cannot
distinguish them. Note that from HTTP standard's perspective, underscore
in the header field is valid, so it may break the application that
intentionally uses underscores.

Envoy returns the response with status code 400 for HTTP/1.1 and returns
RST_STREAM frame with error code INTERNAL_ERROR for HTTP2. The test
cases cover HTTPS+HTTP2 and HTTP+HTTP/1.1.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>