lb: configurable min of healthy backends for successful T1->T2 HC
Currently, the T2 Envoy health check filter is hardcoded to respond
with `200 OK` if a minimum of 20% of the addresses of a backend are
healthy. If the health check filter fails, T1 nodes will quarantine
the endpoints of the given T2 node.

This commit introduces the possibility to configure this percentage
globally for all backends. Be aware that if multiple backends are used
on the same route (e.g. for different routes), all the backends need to be
above their configured minimum value.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>