examples/websocket: Examples for websocket tunneling
Add example resources for websocket tunneling using CiliumEnvoyConfig
CRDs and listener references in CiliumNetworkPolicy.

This example shows how arbitrary TCP traffic can be tunneled via a HTTP
proxy by first encoding it in a WebSocket tunnel and protecting it with
TLS. Tunnel termination is supported with a standalone Envoy container
and also with a CiliumEnvoyConfig within a Cilium cluster. Example
traffic uses ssh and HTTPS.

Instructions are in README.md, and tunnel origination and termination
resources are in 'client' and 'server' subdirectories,
respectively. 'squid' contains the configuration used for the HTTP proxy
used on the origination side.

HTTP proxy redirection uses chained Envoy listeners using the internal
listener feature. This is needed to inject TLS-protected websocket tunnel
into a HTTP CONNECT tunnel for the HTTP proxy.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>