lb: Set default maxConnectionDuration
maxConnectionDuration is an absolute lifetime of the TCP connection.
Regardless of any other timeout, the TCP connection will be closed when
it lives longer than this duration. This is useful when there are many
long-lived connections. In such a case, once the connection is
established with T2 node, it will stick to the same T2 node for a long
time even if the LB admin up/down scale T2 nodes. Same applies for the
upstream.

By default, it is unlimited. That means the TCP connection will never be
closed (unless it hit other timeouts/limits). In this commit, we set
maxConnectionDuration to 1 hour for both downstream and upstream. This
is just a tentative value to let long-lived connections to be rebalanced
at certain frequency instead of sticking to the same host. If this
default doesn't fit for any application, we can expose it as a
configurable parameter later.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>