lb: Add missing enterprise.bgpControlPlane.enabled flag
Currently, the BFD operator only works when
enterprise.bgpControlPlane.enabled is true. Our example and CI setup are
missing this flag, so the BFD was not working at all. Turn on the flag
to make it work.

Note that enterprise.bgpControlPlane.enabled is mutually exclusive with
bgpControlPlane.enabled. That's why we need to disable
bgpControlPlane.enabled here.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>