enterprise/k8s: Add `IsovalentClusterwideEncryptionPolicy` CRD
This commit adds the schema (and auto-generated files) for the new
`IsovalentClusterwideEncryptionPolicy` CRD. This (non-namespaced) CRD
can be used to selectively encrypt traffic between the pods selected by
this new resource. As an example based on the Cilium CLI connectivity
tests, the following policy encrypts all traffic to port 8080/TCP
between all client pods to all echo pods (including reply traffic):

```
apiVersion: isovalent.com/v1alpha1
kind: IsovalentClusterwideEncryptionPolicy
metadata:
  name: encrypt-client-to-echo-8080
spec:
   namespaceSelector:
     matchLabels:
        kubernetes.io/metadata.name: cilium-test
   podSelector:
     matchLabels:
       kind: client
   peers:
   - podSelector:
       matchLabels:
         kind: echo
     namespaceSelector:
       matchLabels:
         kubernetes.io/metadata.name: cilium-test
     ports:
     - protocol: TCP
       port: 8080
```

If both a namespace and pod selector are provided, the selection applies
to selected pods within the selected namespace. If only a namespace
selector is provided, all pods in that namespace are selected.

The CRD validation is intentionally restrictive in this initial version:

  - The namespace selector is always required. This ensures users don't
    accidentally select pods in remote namespaces they weren't meaning
    to. It is still possible to provide an explicit wildcard selector
    (`namespaceSelector: {}`) if the user really wants to select pods
    in all namespaces.
  - At least one explicit peer selector always needs to be provided.
    The datapath implementation does currently not support peer
    wildcarding, thus this is disallowed in the CRD too.
  - At least one explicit port needs to be provided. This restriction
    might be relaxed later (as the datapath implementation should be
    able to support encryption traffic on all ports if requested), but needs
    testing and validation first.

Note that this is an initial version of the CRD. Potential future
additions may include:

  - Adding a `status` field to aid with troubleshooting
  - Allowing port wildcarding (as described above)a
  - Supporting ICMP traffic to be selected
  - Introducing a namespaced variant of the resource
  - Supporting negative rules (aka carve out) selectors

All these additions should be possible without breaking changes in the
current design. For further discussion, please refer to the CFP.

Co-authored-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>