enterprise/maps: Add encryption policy BPF map
This commit adds a new BPF map called `cilium_encryption_policy_map` to
cilium-agent. The map is only created if the encryption policy feature
enabled. A subsequent commit will add the control plane populating the
map based on Kubernetes resources.

The map itself is an LPM trie using a 4-tuple `(subject id, peer id,
proto, port)` as the key and a set of flags for the value. This map will
be used by the datapath (added in a separate commit) to determine if
certain traffic needs to be transparently encrypted.

The commit also adds a basic unit test ensuring accessing the map from
Go works as expected.

Co-authored-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>