enterprise/encryption: Add encryption policy StateDB table
This commit adds a StateDB table which will contain the encryption
policy tuples to be added to the encryption policy BPF map. The StateDB
table primary key is the tuple itself (ensuring there is a 1:1 mapping
between a StateDB and the BPF map). Every entry also tracks its owners
via an `Owners` slice. This slice of `RuleKey`s essentially links it to
rules found the K8s `IsovalentClusterwideEncryptionPolicy` resource.
This will allow us to track when tuples are entries are orphaned and
thus can be removed.

The StateDB entry also contains a status field for the reconciler. This
will be used to reconcile the StateDB data with the BPF map, ensuring
the two are always in sync.

In addition, a `cilium-dgb` subcommand is added to allow the inspection
of the current table state.

Co-authored-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>