bpf: add encryption policy map definition
This commit adds the BPF map needed to realize encryption policies,
which let users select flows between workloads which get encrypted while
other unselected flows stay plain text.

This BPF map stores the following 4-tuple (source identity, destination
identity, protocol, destination port) to track which l3/l4 flows between
identities need to be encrypted. Given we use a LPM Trie map, this allows
us to enable wildcarding of the destination identity, the protocol and
the destination port for this functionality in the future. However, due
to the directionality of the 4-tuple used as a key, it will require two
lookups to identify reply packets that need encryption and to identify
traffic that doesn't need to be encrypted.

Example lookup scenario:

Let's assume the following entry is present in the BPF map:
(10, 20, TCP, 9090).

This policy expresses that TCP traffic from pods with identity 10 to pods
with identity 20 on port 9090 and corresponding reply traffic should be
encrypted. Now Cilium’s datapath sees a TCP packet from a pod with
identity 20 coming from port 9090 to identity 10.

To match this packet the following lookups are needed:

1. (20, 10, TCP, X), lookup current flow direction, no match
2. (10, 20, TCP, 9090), lookup to check the reverse of the connection, match

Signed-off-by: Robin Gögge <r.goegge@isovalent.com>