lb: properly handling portclash on same VIP
Currently, when two `LBFrontend`s are referencing the same `LBVIP`
and define the same port, reconciliation doesn't work as expected.

* Dependent resources are still created
* Status isn't properly set on the `LBFrontend`

The reason is because introducing the LBVIP with #4080 changed the
logic that only the status of the VIP (and indirectly the VIP service)
is checked, but not the LB IPAM status of the actual T1 service. E.g.
a port clash is only visible in the status of the T1 service.

Therefore, this commit reverts part of the logic where the status of the
T1 service from a previous reconciliation needs to be taken into account.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>