lb: Always set ignore_health_on_host_removal
Envoy doesn't stop forwarding to the endpoints removed from the Cluster
when the active health check of them are still alive. Due to this, when
we remove the backend from LBBackendPool, Envoy keeps forwarding when
the health check is still alive. We think the general expectation here
is the backend becomes unreachable immediately when the user removes the
backend from the pool. Set ignore_health_on_host_removal to achieve the
expected behavior.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>