enterprise/encryption: Add encryption policy engine
This commit adds the heart of the encryption policy control-plane.

When the engine observes a new `IsovalentClusterwideEncryptionPolicy` resource,
it will validate and parse that resource (see parse.go), and then add each
subject and peer selector of the policy into a private instance of the
selector cache. The selector cache then provides us with a list of identities
selected by each selector, thus allowing us to compute the required BPF map
tuples which are based on numeric identities.

When the engine observes identity changes, it informs the selector cache about
that change, which in turn will inform us (via `{peer,subjectIdentityNotifier}`)
about the affected policy selectors (and the policy rules the affected selectors
belong to). For every policy rule affected by the identity change, we compute
the set of tuples to insert or remove from the BPF map.

The engine itself does not write directly to the BPF map. Instead, it stores
the computed `EncryptionTuple` in a StateDB table. Every StateDB entry
corresponds to exactly one entry in the BPF map. In the StateDB table,
we keep track of all currently live tuples, as well as the owning policy
rule(s) that generated the tuple.

The actual BPF map itself is updated by a BPF map reconciler that subscribes
to updates to `EncryptionPolicyEntry` and reconciles the StateDB state with the
BPF map accordingly. This reconciler is not part of this commit and will
be added in a subsequent commit.

Co-authored-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>