enterprise/encryption: Add encryption policy BPF map reconciler
This commit adds the last piece of the encryption policy engine, the BPF
map reconciler. This reconciler subscribes to the encryption policy
StateDB table and reconciles it with the encryption policy BPF map.
The reconciler then takes care of updates and deletions on the BPF map
whenever the engine updates the StateDB table.

To ensure that the reconciler does not prune entries before we have
observed all identities and policies, we make use of the StateDB
initializers feature. The BPF reconciler will not prune any stale
entries from the BPF map until the table is initialized. In our case, we
mark the table as initialized once we have observed both a policy and a
identity sync event.

In addition, this commit also provides a `node_config.h` define for the
datapath if the encryption policy feature is enabled. This will ensure
that the datapath is compiled with the encryption policy parts enabled.

Co-authored-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Robin Gögge <r.goegge@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>