bpf: add encryption policy support for wireguard
This commit adds logic to the wireguard hook that extracts necessary
data from the skb to make a lookup to the encryption policy map. For
that lookup we require the security identities of the source and
destination pods, the L4 protocol number and the source and destination
ports.

Initially, this will only be supported for native (IPv4/IPv6) and VXLAN
encapsulated (IPv4/IPv6) TCP and UDP traffic. For native traffic we can
rely on the existing IPcache lookups to fetch the security identities,
and read L4 ports from the usual L4 offsets. For overlay traffic we need
to parse the IPs from the inner packet, and calculate offsets that point
to memory from the inner packet containing L4 port data. As a future
improvement it could be considered to move the lookup to the encryption
policy map to the `to-overlay` program to avoid parsing inner packet
information and mark encapsulated packets with a special mark indicating
the need for encryption which we can read in the wireguard hook.

It should be noted that the way we store encryption policies is
directional. Meaning users will express policies from a request
perspective: (Source, Destination, Protocol, Destination Port). However,
that does not imply only one direction of the connection is being
encrypted. Replies, (Destination, Source, Protocol, Source Port), will
matched as well as the following matching procedure shows:

1. Look up (srcID, dstID, porto, dstPort) in Encryption Policy Map
    → If present, redirect to cilium_wg0
2. Look up (dstID, srcID, porto, srcPort) in Encryption Policy Map
    → If present, redirect to cilium_wg0
3. No match:
    →  egress on native device in clear text

This patch adds BPF integration tests for the following test cases:
- Request match, Native, IPv4, TCP
- Reply match, Native, IPv4, TCP
- No match, Native, IPv4, TCP
- Request match, Native, IPv6, TCP
- Reply match, Native, IPv6, TCP
- No match, Native, IPv6, TCP
- Request match, VXLAN, inner IPv4, TCP
- No match, VXLAN, inner IPv4, TCP
- Request match, VXLAN, inner IPv6, TCP
- No match, VXLAN, inner IPv6, TCP

Signed-off-by: Robin Gögge <r.goegge@isovalent.com>