lb: persistentbackend config on HTTP & TLS passthrough route
This commit introduces support for persistent backend selection. It
can be configured on a route for the application types `HTTPProxy`, `HTTPSProxy`
& `TLSPassthrough`. Depending on the application type, different attributes
are available.

HTTP(S)

* Source IP
* HTTP Cookie
* HTTP Headers

TLS Passthrough

* Source IP

Requests that match a route are sent to the same backend if all of the configured
attributes in the persistent backend configuration are the same.

Notes:

* Persistent backend selection does require the referenced backend to have
LB algorithm `consistentHashing` configured.
* If persistent backend configuration isn't defined and the backend uses LB
  algorithm `consistentHashing`, the backends are "randomly" selected (based on
  how they are spread in the maglev table)

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>