lb: condition for compatible backends on LBService
This commit adds an additional status condition `lb.cilium.io/BackendsCompatible`
on the CRD `LBService`.

If a route of an application (`HTTPProxy`, `HTTPSProxy` or `TLSPassthrough`) configures
persistent backends, the referenced backend needs to have LB algorithm `consistentHashing`
configured. Otherwise the condition is false and the message of the condition contains which
backends aren't compatible.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>