lb: Support TLS Proxy application
Support TLS Proxy application which proxies TLS stream to either TCP or
TLS. It allows SNI-based routing same as the TLS Passthrough application
and TLS configuration same as the one in the HTTPS Proxy application.
When the tlsConfig stanza is defined in the LBBackendPool, it initiates
TLS session on the backend side. Otherwise, it uses TCP connection.

Note that we still need to configure raw_buffer + HTTP Connection
Manager for health checking. The user traffic and health check traffic
are distinguished by the protocol detection with TLS inspector.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>