lb: e2e: remove suite state
Currently, aside from using `TestMain` to bootstrap the "test suite",
the suite also keeps state (clients, T1 client ips) in a global var `suite`.

To avoid this global state, this commit refactors the functions to retrieve
the clients and T1 client ips to avoid having cumbersome error handling that
bloats the test code.

This way we can remove the global suite state and directly retrieve the relevant
info in the tests itself.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>