lb: fix LB IPAM status condition reason propagation
Currently, propagating the LB IPAM status from a K8s Service to
the `LBVIP` fails if there is an unexpected reason.

Reason is that the current logic wraps the reason with `Unexpected (<reason>)`
which results in a reason that isn't allowed by the regex for the
field `^[A-Za-z]([A-Za-z0-9_,:]*[A-Za-z0-9_])?$`.

Therefore, this commit changes the pattern to `unexpected:<reason>`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>