lb: introduce LBBackendPool reconciler
This commit introduces a reconciler that is responsible for the CRD
`LBBackendPool`.

It's primary goal is to provide feedback to the user of the API, whether
the resources has been accepted. Therefore a list of conditions has been
added to the CRD.

In addition, the maglev table size (for loadbalancing algorithm consistent
hashing) is checked to be a prime. If not, the accepted condition is false
and signals the issue with a corresponding reason and message.

Example

```
status:
  conditions:
  - lastTransitionTime: "2024-09-12T12:19:45Z"
    message: 'BackendPool is invalid: .spec.loadBalancing.algorithm.consistentHashing.algorithm.maglev.tableSize
      10 is not prime'
    observedGeneration: 4
    reason: Invalid
    status: "False"
    type: lb.cilium.io/Accepted
```

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>