lb: extract loading existing T1 k8s service into separate function
This commit extracts the attempt to load any existing T1 k8s service into a
separate function. This way it's aligned with the rest of the resources that
are getting loaded during reconciliation.

In addition, this fixes the issue where the t1service reference is never `nil` -
even if not found.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>