lb: remove debug log messages if referenced resources are missing
Currently, the operator still logs (level=debug) if referenced resources
(secrets & backends) don't exist.

As this information is also published on the status of the `LBService`, it's
no longer necessary to log this event (as it's actually not relevant
for the operator of the LB).

Therefore, this commit removes the log messages and the need to pass the
`scopedLogger`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>