egressgwha: track egress interface's ifindex
When a node is selected as gateway by an IEGP, we currently only track the
egress interface's name in the gateway's network config. But with a
subsequent patch we will also make use of the egress interface's ifindex.
Therefore start to also track the ifindex in the gateway's network config.

We currently limit this functionality to IEGPs that either specify the
interface name, or specify *no* interface selector (and thus implicitly
select the interface with the default route).

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>