lb: check for invalid backends when check for compatible backends on service
Currently, the reconciliation of the `LBService` logs a debug message when a
referenced `LBBackendPool` isn't valid (e.g. for maglev table size not being
prime). This isn't the best behavior, as the user of the API that edits the
`LBBackendPool` probably never sees the log of the Cilium Operator (LB
control plane).

With the introduction of a dedicated reconciler for the `LBBackendPool`, we
validate the backend (e.g. prime) and write the status with the condition
accordingly.

Therefore, this commti removes the debug log on the `LBService` and replaces
it with the propagation of the `LBBackendPool` status into the status of the
`LBService`.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>