lb: use 'strip_matching_host_port` on Envoy HCM
Currently, the pseudo HTTP header `:authority` contains the port
if it's provided by the client. This is the case if a port other
than the well-known port (http=80,https=443) is used on the `LBService`.

This leads to challanges matching routes (and later on RBAC) rules - as
one has to differentiate between the different cases.

Therefore, this commit uses the HCM property `strip_matching_host_port`
on the HTTPConnectionManager. This convenience functionality stripts the
port if it's the same as the one configured on the Envoy listener.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>