lb,test: Fix wrong client certificate validation for TLSProxy
In the current TLSProxy test, we provide the TLS certificate to the
LBService.spec.tlsProxy.tlsConfig.validation.secretRef instead of CA
certificate. In this case, Envoy passes all the traffic despite the
requireClientCertitificate is set. Fix it by generating proper CA and
client certificates and provide CA certificate to the secretRef.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>